+++ javascript/node/selenium-webdriver/remote/index.js
@@ -17,22 +17,27 @@
 
 'use strict';
 
-const AdmZip = require('adm-zip'),
-    fs = require('fs'),
-    path = require('path'),
-    url = require('url'),
-    util = require('util');
+const AdmZip = require('adm-zip');
+const fs = require('fs');
+const path = require('path');
+const url = require('url');
+const util = require('util');
 
-const httpUtil = require('../http/util'),
-    io = require('../io'),
-    exec = require('../io/exec'),
-    cmd = require('../lib/command'),
-    input = require('../lib/input'),
-    promise = require('../lib/promise'),
-    webdriver = require('../lib/webdriver'),
-    net = require('../net'),
-    portprober = require('../net/portprober');
+const httpUtil = require('../http/util');
+const io = require('../io');
+const exec = require('../io/exec');
+const cmd = require('../lib/command');
+const input = require('../lib/input');
+const promise = require('../lib/promise');
+const webdriver = require('../lib/webdriver');
+const net = require('../net');
+const portprober = require('../net/portprober');
 
+
+/**
+ * @typedef {(string|!Array<string|number|!stream.Stream|null|undefined>)}
+ */
+var StdIoOptions;
 
 
 /**
@@ -94,8 +99,7 @@ ServiceOptions.prototype.env;
  * IO configuration for the spawned server process. For more information, refer
  * to the documentation of `child_process.spawn`.
  *
- * @type {(string|!Array<string|number|!stream.Stream|null|undefined>|
- *         undefined)}
+ * @type {(StdIoOptions|undefined)}
  * @see https://nodejs.org/dist/latest-v4.x/docs/api/child_process.html#child_process_options_stdio
  */
 ServiceOptions.prototype.stdio;
@@ -286,6 +290,151 @@ class DriverService {
  * @const {number}
  */
 DriverService.DEFAULT_START_TIMEOUT_MS = 30 * 1000;
+
+
+/**
+ * Creates {@link DriverService} objects that manage a WebDriver server in a
+ * child process.
+ */
+DriverService.Builder = class {
+  /**
+   * @param {string} exe Path to the executable to use. This executable must
+   *     accept the `--port` flag for defining the port to start the server on.
+   * @throws {Error} If the provided executable path does not exist.
+   */
+  constructor(exe) {
+    if (!fs.existsSync(exe)) {
+      throw Error(`The specified executable path does not exist: ${exe}`);
+    }
+
+    /** @private @const {string} */
+    this.exe_ = exe;
+
+    /** @private {!ServiceOptions} */
+    this.options_ = {
+      args: [],
+      port: 0,
+      env: null,
+      stdio: 'ignore'
+    };
+  }
+
+  /**
+   * Define additional command line arguments to use when starting the server.
+   *
+   * @param {...string} var_args The arguments to include.
+   * @return {!THIS} A self reference.
+   * @this {THIS}
+   * @template THIS
+   */
+  addArguments(var_args) {
+    let args = Array.prototype.slice.call(arguments, 0);
+    this.options_.args = this.options_.args.concat(args);
+    return this;
+  }
+
+  /**
+   * Sets the host name to access the server on. If specified, the
+   * {@linkplain #setLoopback() loopback} setting will be ignored.
+   *
+   * @param {string} hostname
+   * @return {!DriverService.Builder} A self reference.
+   */
+  setHostname(hostname) {
+    this.options_.hostname = hostname;
+    return this;
+  }
+
+  /**
+   * Sets whether the service should be accessed at this host's loopback
+   * address.
+   *
+   * @param {boolean} loopback
+   * @return {!DriverService.Builder} A self reference.
+   */
+  setLoopback(loopback) {
+    this.options_.loopback = loopback;
+    return this;
+  }
+
+  /**
+   * Sets the base path for WebDriver REST commands (e.g. "/wd/hub").
+   * By default, the driver will accept commands relative to "/".
+   *
+   * @param {?string} basePath The base path to use, or `null` to use the
+   *     default.
+   * @return {!DriverService.Builder} A self reference.
+   */
+  setPath(basePath) {
+    this.options_.path = basePath;
+    return this;
+  }
+
+  /**
+   * Sets the port to start the server on.
+   *
+   * @param {number} port The port to use, or 0 for any free port.
+   * @return {!DriverService.Builder} A self reference.
+   * @throws {Error} If an invalid port is specified.
+   */
+  setPort(port) {
+    if (port < 0) {
+      throw Error(`port must be >= 0: ${port}`);
+    }
+    this.options_.port = port;
+    return this;
+  }
+
+  /**
+   * Defines the environment to start the server under. This setting will be
+   * inherited by every browser session started by the server. By default, the
+   * server will inherit the enviroment of the current process.
+   *
+   * @param {(Map<string, string>|Object<string, string>|null)} env The desired
+   *     environment to use, or `null` if the server should inherit the
+   *     current environment.
+   * @return {!DriverService.Builder} A self reference.
+   */
+  setEnvironment(env) {
+    if (env instanceof Map) {
+      let tmp = {};
+      env.forEach((value, key) => tmp[key] = value);
+      env = tmp;
+    }
+    this.options_.env = env;
+    return this;
+  }
+
+  /**
+   * IO configuration for the spawned server process. For more information,
+   * refer to the documentation of `child_process.spawn`.
+   *
+   * @param {StdIoOptions} config The desired IO configuration.
+   * @return {!DriverService.Builder} A self reference.
+   * @see https://nodejs.org/dist/latest-v4.x/docs/api/child_process.html#child_process_options_stdio
+   */
+  setStdio(config) {
+    this.options_.stdio = config;
+    return this;
+  }
+
+  /**
+   * Creates a new DriverService using this instance's current configuration.
+   *
+   * @return {!DriverService} A new driver service.
+   */
+  build() {
+    let port = this.options_.port || portprober.findFreePort();
+    let args = Promise.resolve(port).then(port => {
+      return this.options_.args.concat('--port=' + port);
+    });
+
+    let options =
+        /** @type {!ServiceOptions} */
+        (Object.assign({}, this.options_, {args, port}));
+    return new DriverService(this.exe_, options);
+  }
+};
 
 
 /**